andycot v0.0.1 Andycot.AuctionProcessor

Functions

handle_cast(arg, state)

Inserts the BidPlaced event in the event store When set to true the field from_migration allows to furnish an already defined created_at date, otherwise the created_at date is filled with the current date of insertion

init(initial_value)

start_link(initial_value)