andycot v0.0.1 Supervisor.Auction

Functions

cancel_auction(info, mode)

Cancels an auction

Returns: { :ack, :auction_cancelled } upon successfull cancelation { :nack, :not_owner_or_system } only the seller or the system (0) are allowed to cancel an auction { :nack, :too_late } an auction having bids can only be cancelled before a certain time limit

close_auction(info, mode)

Closes an auction

Returns : { :nack, :has_bids } { :nack, :not_owner_or_system }

convert_string_key_to_atom_key(the_map)

create_auction(command, mode \\ :standard)

Creates an auction and starts a worker for this auction, If the auction_id given in the command is nil then a new auction_id is allocated

Returns the created auction { :ok, %Worker.Auction{} }

duplicate_auction(command, mode \\ :standard)

Duplicates an auction and starts a worker for this new auction,

Returns the created auction { :ok, %Worker.Auction{} }

get_auction(auction_id, mode \\ :standard)

Returns the auction state given an auction_id The return type is %Worker.Auction{}

get_chunk_of_events(page_number)

init()

place_bid(bid, mode \\ :standard)

Successfull returns { :ack, :bid_placed } (for variable price auctions) { :ack, :bid_placed_time_extended } (for variable price auctions) { :ack, :bid_placed_auction_will_close } (for fixed price auctions)

Failed returns { :nack, :auction_has_ended } { :nack, :auction_is_suspended } { :nack, :self_bidding } { :nack, :auction_has_ended } { :nack, :auction_not_yet_started } { :nack, :wrong_requested_qty } { :nack, :not_enough_stock } { :nack, :bid_below_allowed_min } { :nack, :wrong_requested_qty } (for fixed price auctions) { :nack, :wrong_bid_price } (for fixed price auctions)

start_all_workers(page_number \\ 1)

start_chunk_of_workers(page_number, events)

start_link()

start_worker(auction, mode)