andycot v0.0.1 Worker.Auction

A worker to manage the state and persistence of an auction using gen_server

It works in three different modes :

* `:legacy` used during the migration phase (andycot v1 to v2)
* `:replay` used when restarting the system, created the workers and replays the events
* `:standard` used during the as-asual web-site activity

It handles and generates the following Commands and Events

:create_auction_command             :auction_created_event
:place_bid_command                      :bid_placed_event
:close_auction_command              :auction_closed_event

The following commands can be issued for testing purposes

Sale type 1 with automatic_renewal = true
    Autogenerate auction_id and place a bid
    Supervisor.Auction.create_auction(%Command.CreateAuction{seller_id: 269, sale_type_id: 1, listed_time_id: 1, start_date_time: Andycot.Tools.Timestamp.now(), end_date_time: Andycot.Tools.Timestamp.now()+10, start_price: 1.00, bid_up: 0.10, stock: 1}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 260, requested_qty: 1, max_value: 1.00, created_at: Andycot.Tools.Timestamp.now()}, :standard)

Sale type 1 with automatic_renewal = false and stock = 1 and reserve_price = nil
    Autogenerate auction_id and place a bid
    Supervisor.Auction.create_auction(%Command.CreateAuction{seller_id: 269, sale_type_id: 1, listed_time_id: 1, start_date_time: Andycot.Tools.Timestamp.now(), end_date_time: Andycot.Tools.Timestamp.now()+240, start_price: 1.00, bid_up: 0.10, stock: 1}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 260, requested_qty: 1, max_value: 2.00, created_at: Andycot.Tools.Timestamp.now()}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 261, requested_qty: 1, max_value: 1.30, created_at: Andycot.Tools.Timestamp.now()}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 260, requested_qty: 1, max_value: 3.00, created_at: Andycot.Tools.Timestamp.now()}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 262, requested_qty: 1, max_value: 2.50, created_at: Andycot.Tools.Timestamp.now()}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 261, requested_qty: 1, max_value: 3.00, created_at: Andycot.Tools.Timestamp.now()}, :standard)

Sale type 1 with automatic_renewal = true and stock = 4 and reserve_price = nil
    Autogenerate auction_id and place a bid
    Supervisor.Auction.create_auction(%Command.CreateAuction{seller_id: 269, sale_type_id: 1, listed_time_id: 1, start_date_time: Andycot.Tools.Timestamp.now(), end_date_time: Andycot.Tools.Timestamp.now()+20, start_price: 1.00, bid_up: 0.10, stock: 4}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 260, requested_qty: 1, max_value: 1.40, created_at: Andycot.Tools.Timestamp.now()}, :standard)

Sale type 1 with automatic_renewal = true and reserve_price not met
    Supervisor.Auction.create_auction(%Command.CreateAuction{seller_id: 269, sale_type_id: 1, listed_time_id: 1, start_date_time: Andycot.Tools.Timestamp.now(), end_date_time: Andycot.Tools.Timestamp.now()+30, start_price: 1.00, bid_up: 0.10, stock: 1, reserve_price: 2.00}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 260, requested_qty: 1, max_value: 1.80, created_at: Andycot.Tools.Timestamp.now()}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 262, requested_qty: 1, max_value: 1.90, created_at: Andycot.Tools.Timestamp.now()}, :standard)

Sale type 1 with automatic_renewal = false
    Supervisor.Auction.start_worker(%Worker.Auction{auction_id: 100, seller_id: 269, sale_type_id: 1, listed_time_id: 1, is_closed: false, start_date_time: Andycot.Tools.Timestamp.now(), end_date_time: Andycot.Tools.Timestamp.now()+30, is_suspended: false, start_price: 1.00, current_price: 1.00, bid_up: 0.10, stock: 1, type_id: 1, automatic_renewal: false}, :standard)

Sale type 2 with automatic_renewal = true
    Supervisor.Auction.create_auction(%Command.CreateAuction{seller_id: 269, sale_type_id: 2, listed_time_id: 1, start_date_time: Andycot.Tools.Timestamp.now(), end_date_time: Andycot.Tools.Timestamp.now()+10, start_price: 1.00, bid_up: 0.10, stock: 1}, :standard)
    Supervisor.Auction.place_bid(%Command.PlaceBid{auction_id: 1, bidder_id: 260, requested_qty: 1, max_value: 1.00, created_at: Andycot.Tools.Timestamp.now()}, :standard)

Sale type 2 with automatic_renewal = false
    Supervisor.Auction.start_worker(%Worker.Auction{auction_id: 100, seller_id: 269, sale_type_id: 2, listed_time_id: 1, is_closed: false, start_date_time: Andycot.Tools.Timestamp.now(), end_date_time: Andycot.Tools.Timestamp.now()+30, is_suspended: false, start_price: 1.00, current_price: 1.00, bid_up: 0.10, stock: 10, automatic_renewal: false, type_id: 1}, :standard)

Functions

add_amounts(value, bid_up)

apply_event(event, mode, auction_state)

Process the AuctionSold event

Returns: { :ack, :auction_sold}

can_be_ended_early?(auction_state)

clear_ticker(auction_state)

compute_end_date_time(start_date_time, listed_time_id)

convert_string_key_to_atom_key(the_map)

get_days_given_listed_time_id(listed_time_id)

handle_call(msg, from, state)

Used for debug purposes

handle_info(arg1, state)

Handles the ticker message used to check if an auction of fixed price has ended and update its status, eventually sending emails to the buyer and the seller, updating the quote …

init(args)

make_auction_cancelled_event(command, overwrite)

make_auction_closed_event(command, overwrite)

make_auction_created_event(command, overwrite)

make_auction_renewed_event(command, overwrite)

make_auction_sold_event(command, overwrite)

make_bid_from_event(bid_placed, overwrite)

make_bid_placed_event(command, overwrite)

make_bid_rejected_event(command, overwrite)

make_cancel_rejected_event(command, overwrite)

make_close_rejected_event(command, overwrite)

make_worker_id(auction_id)

Generates a unique worker id

map_merge_keep_right(k, vl, vr)

Helper function used when merging two maps where we want to keep the values of the map given as the right parameter

maybe_extend_time(auction_state)

normalize_command(command, bid_up)

normalize_value(value, bid_up)

process_command(command, mode, auction_state)

Process the CancelAuction command

Returns: { :nack, :not_owner_or_system} only the seller or the system (0) are allowed to cancel an auction { :nack, :too_late} an auction having bids can only be cancelled before a certain time limit

{ :ack, :auction_cancelled} upon successfull cancelation

process_ticker(auction_state)

Process the ticker message on a fixed price auction

Returns : { :nack, :auction_has_not_ended} { :nack, :auction_cancelled}

{ :ack, :auction_closed_no_bids_no_automatic_renewal} { :ack, :auction_sold} { :ack, :renew_auction}

replay_events(auction_state)

Replays all events except the AuctionCreated event Used in the :replay mode

replay_next_event(list, auction_state)

Handles the replay of the next event Helps to stop the recursion when the list of events to replay is empty

reply_maybe_hibernate(auction_state, reply)

Handles the hibernation of our auction processes depending on the remaining time until the auction’s end Goes to hibernation if the remaining time is more than 1 hour

start_link(auction, mode)

start_ticker(auction, pid \\ nil)

Schedules a timer that will fire at the auctions’s end time

Returns the given auction updated with the timer ref or the original auction if the time could not be started

stop_ticker(auction_state)

update_auction_state(auction_state, new_current_price, new_bids)

via_tuple(auction_id)

Generates a unique worker name